==> scripts/legacy/DeisyCode/03_Differential_Net_Analysis.Rmd <==
---
title: "Find Genes Differentially co-expressed"
output:
  html_document:
    theme: united
    highlight: tang
author: Deisy Morselli Gysi
---
#### Northeastern University
#### Project: Scipher RA - Responder vs Non-responders to TNF drugs

## Description
The aim here is to identify Genes Differentially Co-expressed, for that we will use CoDiNA. 


```{r, message=FALSE}
require(dplyr)
require(magrittr)
require(data.table)
require(ggplot2)
require(CoDiNA)
# Load CN

nR = fread(
  "../data/out/wTO/NonResponder_CN.csv")
R = fread(
  "../data/out/wTO/Responder_CN.csv")
Genes = c(nR$Node.1, nR$Node.2, R$Node.1, R$Node.2) %>% unique()
Genes = data.frame(Node.1 = Genes,
                   Node.2  = Genes,
                   CN = 0,
                   pval.fisher = 0)

# Make sure all genes that were measured have a shot at the analysis, otherwise CoDiNA will remove them. 
# 
nR = rbind(nR, Genes)
R = rbind(R, Genes)

# We will use only interactions that also exist in the PPI
# 
PPI = fread("~/Desktop/PostDoc/02_Toolbox/03_Data/PPI_Symbol_Entrez.csv")
PPI = PPI[,-c(1,2)] %>% unique()
PPI %<>%
  filter(Symbol_A != "" & Symbol_B != "") %>%
  filter(Symbol_A != Symbol_B)

PPI$PPI = 1
PPI = OrderNames(PPI) %>% unique() #remove duplicated links
```


```{r}
require(CoDiNA)
Diff = MakeDiffNet(Data = list(PPI, R, nR), 
                   Code = c('PPI', "R", "nR"), 
                   stretch = F) # stretch is used when the networks come from different studies, which is not the case here.
Diff_cl = Diff %>% filter(Score_ratio > 1) # ## Select only strong and well classified edges
## 
## Compare non_filtered vs filtered
Diff_cl %>% group_by(Phi_tilde) %>% summarize(n = n())
Diff %>% group_by(Phi_tilde) %>% summarize(n = n())

# Select only edges that belongs in the PPI
Diff_cl_PPI = Diff_cl %>% filter(Phi_tilde %in% c("a", "g.PPI.nR", "g.PPI.R"))

fwrite(Diff_cl_PPI, '../data/out/CoDINA/CoDiNA_PPI_CL.csv')

Nodes = ClusterNodes(DiffNet = Diff_cl_PPI, cutoff.external = 0, cutoff.internal = 1)
table(Nodes$Phi_tilde)

Node_Cat = Nodes %>%  #filter nodes bt fdr corrected pvalue, for node classification
  mutate(p_adj = p.adjust(pval_Phi_Tilde, method = 'fdr')) %>%
  filter(p_adj < 0.01) %>%
  select(Node, Phi_tilde)


Nodes %>% 
  mutate(p_adj = p.adjust(pval_Phi_Tilde, method = 'fdr')) %>% 
  filter(p_adj < 0.01) %>% 
  group_by(Phi_tilde) %>%
  summarise(n = n())

Node_Cat = Nodes %>% 
  mutate(p_adj = p.adjust(pval_Phi_Tilde, method = 'fdr')) %>% 
  filter(p_adj < 0.01) %>%
  select(Node, Phi_tilde)

Node_Cat %>% 
  filter(Phi_tilde == "g.PPI.R") %>% 
  pull(Node) %>%
  write.table("../data/out/CoDINA/CN_R.txt", quote = F, 
              row.names = F, col.names = F)

Node_Cat %>% 
  filter(Phi_tilde == "g.PPI.nR") %>% 
  pull(Node) %>%
  write.table("../data/out/CoDINA/CN_nR.txt", quote = F, 
              row.names = F, col.names = F)

Node_Cat %>% 
  filter(Phi_tilde == "a") %>% 
  pull(Node) %>%
  write.table("../data/out/CoDINA/CN_a.txt", quote = F, 
              row.names = F, col.names = F)


Node_Cat %>% 
  filter(Phi_tilde != "U") %>% 
  # pull(Node) %>%
  write.table("../data/out/CoDINA/CN_All.txt", quote = F, 
              row.names = F, col.names = F)
```
## Session Info 
```{r}
sessioninfo::session_info()
```
